fix: refactor zarm-icon and fix build error in vite (#834)
