+++ packages/zarm-icons/src/icon/index.tsx
@@ -1,6 +1,6 @@
-import React, { PureComponent, HTMLAttributes } from 'react';
+import React, { HTMLAttributes } from 'react';
 import classnames from 'classnames/dedupe';
-import BasePropsType from './PropsType';
+import BasePropsType from './interface';
 import createFromIconfont from './IconFont';
 
 const INNER_SVG_PROPS = {
@@ -15,71 +15,72 @@ export type IconProps = {
 } & BasePropsType &
   HTMLAttributes<HTMLElement>;
 
-class Icon extends PureComponent<IconProps, {}> {
-  static displayName = 'Icon';
+interface CompoundedComponent
+  extends React.ForwardRefExoticComponent<IconProps & React.RefAttributes<HTMLElement>> {
+  createFromIconfont: typeof createFromIconfont;
+}
 
-  static defaultProps: IconProps = {
-    prefixCls: 'za-icon',
-    theme: 'default',
-    size: 'md',
-    viewBox: INNER_SVG_PROPS.viewBox,
+const Icon = React.forwardRef<HTMLElement, IconProps>((props, ref) => {
+  const {
+    className,
+    type,
+    style,
+    prefixCls,
+    theme,
+    size,
+    children,
+    component: SvgComponent,
+    viewBox,
+    ...rest
+  } = props;
+  const cls = classnames(prefixCls, className, {
+    [`${prefixCls}--${type}`]: !!type,
+    [`${prefixCls}--${theme}`]: !!theme,
+    [`${prefixCls}--${size}`]: !!size,
+  });
+
+  const newProps = {
+    className: cls,
+    style,
+    type,
+    theme,
+    size,
+    ...rest,
   };
 
-  static createFromIconfont = createFromIconfont;
-
-  render() {
-    const {
-      className,
-      type,
-      style,
-      prefixCls,
-      theme,
-      size,
-      children,
-      component: SvgComponent,
-      viewBox,
-      ...rest
-    } = this.props;
-    const cls = classnames(prefixCls, className, {
-      [`${prefixCls}--${type}`]: !!type,
-      [`${prefixCls}--${theme}`]: !!theme,
-      [`${prefixCls}--${size}`]: !!size,
-    });
-
-    const newProps = {
-      className: cls,
-      style,
-      type,
-      theme,
-      size,
-      ...rest,
-    };
-
-    // svg component > children by iconfont > type
-    if (SvgComponent) {
-      INNER_SVG_PROPS.viewBox = viewBox!;
-      return (
-        <i {...newProps}>
-          <SvgComponent {...INNER_SVG_PROPS}>{children}</SvgComponent>
-        </i>
-      );
-    }
-
-    if (children) {
-      return (
-        <i {...newProps}>
-          <svg {...INNER_SVG_PROPS}>{children}</svg>
-        </i>
-      );
-    }
-
-    if (type) {
-      const MyIcon = Icon.createFromIconfont('//at.alicdn.com/t/font_1285036_9ze0vm44z57.js'); // generated by iconfont.cn
-      return <MyIcon {...newProps} />;
-    }
-
-    return null;
+  // svg component > children by iconfont > type
+  if (SvgComponent) {
+    INNER_SVG_PROPS.viewBox = viewBox!;
+    return (
+      <i {...newProps} ref={ref}>
+        <SvgComponent {...INNER_SVG_PROPS}>{children}</SvgComponent>
+      </i>
+    );
   }
-}
+
+  if (children) {
+    return (
+      <i {...newProps} ref={ref}>
+        <svg {...INNER_SVG_PROPS}>{children}</svg>
+      </i>
+    );
+  }
+
+  if (type) {
+    const MyIcon = createFromIconfont('//at.alicdn.com/t/font_1285036_9ze0vm44z57.js'); // generated by iconfont.cn
+    return <MyIcon {...newProps} />;
+  }
+  return null;
+}) as CompoundedComponent;
+
+// Icon.createFromIconfont = createFromIconfont;
+
+Icon.displayName = 'Icon';
+Icon.defaultProps = {
+  prefixCls: 'za-icon',
+  theme: 'default',
+  size: 'md',
+  viewBox: INNER_SVG_PROPS.viewBox,
+};
 
 export default Icon;
